+++ src/ui-dialog/ui-dialog.d.ts
@@ -27,6 +27,7 @@ export class UIDialog<T extends UIDialog_Events> extends HTMLElement {
         events: Events<T>;
         title: string;
         fullscreen: boolean;
+        nofooter: boolean;
         /**
          * @param {boolean} modal
          * @param {boolean} [inert] - This will prevent the autofocus on input elements (default: true)